Add OmniSearch UI
Add a new search kind of search, replacing the new current catalog
search. OmniSearch can search users, projects, and definitions. It does
so by using 3 different search endpoints: "Entity Search" (for users and
projects), "Definition Search", and "Name Search" (for autocompleting
names/tokens mentioned in definitions, ex: "Remote" or "List").

---

This is currently in a "mocked state", meaning that "definition search"
and "name search" are not hooked up to real endpoint. That is a separate
effort once the APIs for those are ready