Setup Netlify Edge Function for social content
Set up a Netlify Edge Function that replaces the main template with
content from the Share API matching the URL. This helps unfurls of URLs
that doesn't run the client side, such as sharing links on social media
or chat applications like Slack and Discord.

This sets up the replacement infrastructure and replaces the content for
user pages.