Enable contribution diffs
Enable the first version (albeit a bit rough) of contribution diffs.
When navigating to changes of a contribution, show a change overview in a
sidebar and all the changed definitions in a main pane. Expanding each
item will fetch the appropriate diff (or just the definition if it was
an add or remove). Hovering over items in the diff will show more
details about the items.