Disable edit and delete for non-commenters
The frontend supported the project maintainers to be able to edit and
delete any comments on their projects as a form of moderation (similar
to GitHub's behavior). Currently the backend does not support this and
the buttons don't work, so for now disable them until we're ready to
support moderation in full.