Fix CSS variable checker breaking build
For a while now a false negative would occur with the css variable
checker (which goal is to ensure we don't reference undefined css
variables). Turns out, in omni-search.css one of the
var(--font-size-medium), even though it looked correct and with the
right characters would not pass a JavaScript equality check thus
resulting in the false negative. I just rewrote that line to fix the
problem. Even the diff says that they are different, but offer no
indication of what is different. :shrug: