Guard against how Browsers encode spaces
When using the Browser's address bar to perform a search on Share it
uses `+` instead of `%20` to encode spaces and this ends up as literal
`+` when parsed with `Url.percentDecode`. To solve this, replace all `+`
with `%20%` before calling `Url.percentDecode`.

Note: this will still work for search queries where the user wants to
find "+", which is encoded by `Url.percentEncode` as `%2b`.