Add `NewOrgModal`
Add a new app modal, that allows users to create organizations.
It supports suggesting handles from the name and checking if handles are
taken or not. It also allows the user to pick between public (free)
orgs, and commercial (paid) orgs. When picking a commercial org, the
backend will create a support ticket (in addition to the org).