Project: Don't show the Finder with an open modal
On the project overview page, while having a modal open--say the edit
summary modal--hitting the keyboard shortcut (/) to open the Finder (for
searching definitions) would dismiss the active modal and cause the user
to loose their current work and flow.

Disable this behavior by simply checking if a modal is open or not.