Use UCM Desktop Workspace with dependents support
Port the "Workspace" modules from UCM Desktop to Share (eventually
planned to be fully shared via ui-core), and with it the ability to
render other items in the Workspace than just definitions. The first of
which is a new Dependents card that lists the dependents of a
definition.

This includes a considerable amount of configurable modules to allow for
both UCM Desktop and Share to have their unique configuration of
features (though they are possible going to get closer in time).