Add support for diff lines
The API now sends back diffs on a line by line basis with spacers such
that left and right side diffs can be aligned. Update the JSON decoders
to support this new format and render the UI appropriately.

Also add line numbers and change indicators in a diff gutter as well as
line spacer patterns.